QtImGui & thedmd's node editor's "blueprint example"
thedmd's "blueprint" example is more developed than the basic one we implemented last time.  In this change, we needed to start supporting texture loading, which is done in texture_manager.  We also needed to load a font into ImGui. Font loading is not in a library code - it's supposed to be in the in the platform/ host code.  So, it's implemented in the NodosWidget::intializeGL().